CarRent/requirements.txt
asgiref==3.4.0
boto3==1.17.105
botocore==1.20.105
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==3.2.4
django-cors-headers==3.10.0
django-filter==21.1
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
gunicorn==20.1.0
httpie==2.5.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.3.0
psycopg2-binary==2.9.3
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
